feat(cost estimation): updates the cost estimation after adding the extra tags PE-2365
