+++ lib/entities/entity.dart
@@ -1,11 +1,10 @@
 import 'package:ardrive/services/services.dart';
+import 'package:ardrive/utils/app_platform.dart';
 import 'package:arweave/arweave.dart';
 import 'package:cryptography/cryptography.dart';
-import 'package:device_info_plus/device_info_plus.dart';
 import 'package:flutter/foundation.dart';
 import 'package:json_annotation/json_annotation.dart';
 import 'package:package_info_plus/package_info_plus.dart';
-import 'package:platform/platform.dart';
 
 import 'entities.dart';
 
@@ -86,7 +85,7 @@ extension TransactionUtils on TransactionBase {
     DateTime? unixTime,
   }) async {
     addTag(EntityTag.appName, 'ArDrive-App');
-    addTag(EntityTag.appPlatform, _platform);
+    addTag(EntityTag.appPlatform, getPlatform());
 
     // TODO: PE-2380
     // addTag(EntityTag.appPlatformVersion, await _platformVersion);
@@ -101,57 +100,5 @@ extension TransactionUtils on TransactionBase {
   /// Tags this transaction with the ArFS version currently in use.
   void addArFsTag() {
     addTag(EntityTag.arFs, '0.11');
-  }
-
-  // TODO: PE-2380
-  get platformVersion async {
-    final platform = _platform;
-    final deviceInfoPlugin = DeviceInfoPlugin();
-
-    switch (platform) {
-      case 'Android':
-        final androidDeviceInfo = await deviceInfoPlugin.androidInfo;
-        final String? androidVersion = androidDeviceInfo.version.release;
-        final int? sdkVersion = androidDeviceInfo.version.sdkInt;
-        final versionString = '$androidVersion (SDK $sdkVersion)';
-
-        return versionString;
-
-      case 'iOS':
-        final iosDeviceInfo = await deviceInfoPlugin.iosInfo;
-        final String? systemName = iosDeviceInfo.systemName;
-        final String? iosVersion = iosDeviceInfo.systemVersion;
-        final versionString = '$systemName $iosVersion';
-
-        return versionString;
-
-      default: // case 'Web':
-        final webDeviceInfo = await deviceInfoPlugin.webBrowserInfo;
-        final browserName = describeEnum(webDeviceInfo.browserName);
-        final browserVersion = webDeviceInfo.appVersion;
-        final versionString = '$browserName $browserVersion';
-
-        return versionString;
-    }
-  }
-
-  get _platform {
-    if (kIsWeb) {
-      return 'Web';
-    }
-
-    const platform = LocalPlatform();
-
-    /// A string (linux, macos, windows, android, ios, or fuchsia) representing the operating system.
-    final String operatingSystem = platform.operatingSystem;
-
-    switch (operatingSystem) {
-      case 'android':
-        return 'Android';
-      case 'ios':
-        return 'iOS';
-      default:
-        throw Exception('Unsupported platform $operatingSystem!');
-    }
   }
 }
