+++ lib/utils/bundles/fake_tags.dart
@@ -1,5 +1,7 @@
 import 'package:ardrive/entities/entities.dart';
+import 'package:ardrive/utils/app_platform.dart';
 import 'package:arweave/arweave.dart';
+import 'package:package_info_plus/package_info_plus.dart';
 
 final fakePrivateTags = [
   Tag(EntityTag.contentType, ContentType.octetStream),
@@ -10,12 +12,23 @@ final fakePrivateTags = [
   ),
 ];
 
-final fakeApplicationTags = [
-  Tag(EntityTag.appName, 'ArDrive-Web'),
-  Tag(EntityTag.appVersion, '0.0.0'),
-  Tag(EntityTag.unixTime,
-      (DateTime.now().millisecondsSinceEpoch ~/ 1000).toString())
-];
+fakeApplicationTags() async {
+  final packageInfo = await PackageInfo.fromPlatform();
+  final version = packageInfo.version;
+
+  return <Tag>[
+    Tag(EntityTag.appName, 'ArDrive-App'),
+    Tag(EntityTag.appVersion, version),
+    Tag(
+      EntityTag.unixTime,
+      (DateTime.now().millisecondsSinceEpoch ~/ 1000).toString(),
+    ),
+    Tag(EntityTag.appPlatform, getPlatform()),
+
+    // TODO: PE-2380
+    // Tag(EntityTag.appPlatformVersion, getPlatformVersion()),
+  ];
+}
 
 List<Tag> createFakeEntityTags(FileEntity entity) => <Tag>[
       Tag(EntityTag.arFs, '0.11'),
